修复监听非法描述符出错的BUG
1. 修复监听非法描述符出错的BUG;
2.

+++ src/session.c
@@ -513,7 +513,8 @@ int32_t _append_session( struct hashtable * table, struct session * s )
 
 	if ( node->session != NULL && node->session->id == s->id )
 	{
-		syslog(LOG_WARNING, "%s(Index=%d): the SID (Seq=%u, Sid=%ld) conflict !", __FUNCTION__, SID_INDEX(s->id), SID_SEQ(s->id), s->id );
+		syslog(LOG_WARNING, "%s(Index=%d): the SID (Seq=%u, Sid=%ld) conflict !",
+                __FUNCTION__, (int32_t)SID_INDEX(s->id), (uint32_t)SID_SEQ(s->id), s->id );
 		return -2;
 	}
